flask-server/requirements.txt
bcrypt==4.0.1
bleach==6.0.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
Deprecated==1.2.14
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.4.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
psycopg2-binary==2.9.7
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
rich==13.5.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.0.1
